+++ src/main/java/com/pardot/rhombus/cobject/statement/UnboundableCQLStatementIterator.java
@@ -90,7 +90,7 @@ public class UnboundableCQLStatementIterator extends BaseCQLStatementIterator {
 	}
 
 	@Override
-	public void setLimit(int limit){
+	public void setLimit(long limit){
 		this.numberRemaining = limit;
 	}
